Change logger location
